# docs/apifox-openapi.yaml
openapi: 3.0.3
info:
  title: 电影票后端联调 API
  version: 0.2.0
  license:
    name: Proprietary
  description: |
    闲鱼插件、微信机器人、管理端与良票上游联调接口。

    鉴权分为三组：闲鱼接口使用 `X-Plugin-Token`，微信机器人接口使用
    `X-Bot-Token`，报价、订单和上游管理接口使用 `X-Admin-Token`。

    带有“高风险”字样的接口可能创建真实订单、提交上游或支付。手动测试前请确认
    `ticket.upstream.mock-enabled` 与 `ticket.xianyu.auto-fulfillment-enabled` 的配置。
servers:
  - url: http://127.0.0.1:8080
    description: 本地开发环境
...
paths:
...
  /api/orders:
    post:
      tags: [3. 订单管理（高风险）]
      summary: 【高风险】根据报价创建并异步提交订单
      description: 创建本地订单后可能异步提交真实上游并支付。仅在明确确认真实出票后调用。
      operationId: createOrder
      security:
        - adminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              quoteNo: '{{quoteNo}}'
              customerId: apifox-customer-001
              paymentNo: apifox-payment-001
      responses:
        '200':
          $ref: '#/components/responses/OrderOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/AuthNotConfigured'
  /api/orders/{orderNo}:
    get:
      tags: [3. 订单管理（高风险）]
      summary: 查询本地订单
      operationId: getOrder
      security:
        - adminToken: []
      parameters:
        - $ref: '#/components/parameters/OrderNo'
      responses:
        '200':
          $ref: '#/components/responses/OrderOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/AuthNotConfigured'
  /api/orders/{orderNo}/submit/retry:
    post:
      tags: [3. 订单管理（高风险）]
      summary: 【高风险】安全续跑已有上游订单
      description: 仅续跑已经存在明确上游订单号的订单，不应创建第二张上游订单；仍可能执行支付。
      operationId: retryOrderSubmit
      security:
        - adminToken: []
      parameters:
        - $ref: '#/components/parameters/OrderNo'
      responses:
        '200':
          $ref: '#/components/responses/OrderOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/AuthNotConfigured'
  /api/orders/{orderNo}/submit/repeat:
    post:
      tags: [3. 订单管理（高风险）]
      summary: 【极高风险】强制重新提交上游订单
      description: 默认拒绝执行。只有 `force=true` 才会强制重新提交，可能造成重复订单和重复支付。
      operationId: repeatOrderSubmit
      security:
        - adminToken: []
      parameters:
        - $ref: '#/components/parameters/OrderNo'
        - name: force
          in: query
          required: false
          description: 必须显式设为 true 才允许重新提交；可能重复下单。
          schema:
            type: boolean
            default: false
          example: false
      responses:
        '200':
          $ref: '#/components/responses/OrderOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/AuthNotConfigured'
  /api/orders/{orderNo}/sync:
    post:
      tags: [3. 订单管理（高风险）]
      summary: 从上游同步订单状态
      description: 查询已有上游订单状态，不创建新订单。
      operationId: syncOrder
      security:
        - adminToken: []
      parameters:
        - $ref: '#/components/parameters/OrderNo'
      responses:
        '200':
          $ref: '#/components/responses/OrderOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/AuthNotConfigured'
...
  /api/xianyu/orders/{platformOrderId}:
    get:
      tags: [4. 闲鱼插件]
      summary: 查询单个闲鱼订单
      operationId: getXianyuOrder
      security:
        - xianyuPluginToken: []
      parameters:
        - $ref: '#/components/parameters/PlatformOrderId'
      responses:
        '200':
          $ref: '#/components/responses/XianyuOrderOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/AuthNotConfigured'
...
  /api/bot/messages/image:
    post:
      tags: [5. 微信机器人]
      summary: 微信私聊图片报价
      description: '`wechatId + messageId` 为幂等键；重试同一消息必须保持 `messageId` 不变。'
      operationId: createBotQuoteFromImage
      security:
        - botToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BotImageQuoteRequest'
            example:
              wechatId: wx_apifox_001
              nickname: ApiFox 测试用户
              avatarUrl: ''
              messageId: wx-message-001
              imageBase64: '{{ticketImageBase64}}'
      responses:
        '200':
          $ref: '#/components/responses/BotQuoteOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/AuthNotConfigured'
  /api/bot/payments/confirm:
    post:
      tags: [5. 微信机器人]
      summary: 确认微信用户收款
      description: '`paymentNo` 为幂等键；重复请求必须保持用户、报价和金额一致。'
      operationId: confirmBotPayment
      security:
        - botToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BotPaymentConfirmRequest'
            example:
              wechatId: wx_apifox_001
              quoteNo: '{{quoteNo}}'
              amount: 58.38
              paymentNo: wx-payment-001
              proofImageUrl: ''
              confirmer: apifox
              remark: 手工测试收款确认
      responses:
        '200':
          $ref: '#/components/responses/BotPaymentOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/AuthNotConfigured'
  /api/bot/orders:
    post:
      tags: [5. 微信机器人]
      summary: 【高风险】收款后创建并提交订单
      description: 校验报价和收款记录后创建本地订单，并可能异步提交、支付真实上游。
      operationId: createBotOrder
      security:
        - botToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BotCreateOrderRequest'
            example:
              wechatId: wx_apifox_001
              quoteNo: '{{quoteNo}}'
              paymentRecordNo: '{{paymentRecordNo}}'
      responses:
        '200':
          $ref: '#/components/responses/OrderOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/AuthNotConfigured'
  /api/bot/orders/{orderNo}:
    get:
      tags: [5. 微信机器人]
      summary: 微信机器人轮询订单
      description: '`shouldPoll=false` 时停止轮询；`status=ISSUED` 时读取 `ticketCodeInfo`。'
      operationId: getBotOrder
      security:
        - botToken: []
      parameters:
        - $ref: '#/components/parameters/OrderNo'
      responses:
        '200':
          $ref: '#/components/responses/OrderOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/AuthNotConfigured'
components:
  securitySchemes:
    adminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: 本地 profile 默认值为 `local-admin-token`
    xianyuPluginToken:
      type: apiKey
      in: header
      name: X-Plugin-Token
      description: 用户控制台创建且已由管理员设置有效期的插件 Token；请求还必须携带 X-Plugin-Installation-Id
    botToken:
      type: apiKey
      in: header
      name: X-Bot-Token
      description: 用户控制台创建且已由管理员设置有效期的机器人 Token；请求还必须携带 X-Bot-Installation-Id
  parameters:
...
    OrderNo:
      name: orderNo
      in: path
      required: true
      schema:
        type: string
      example: O20260516120000ABCDEFGH
...
    PlatformOrderId:
      name: platformOrderId
      in: path
      required: true
      schema:
        type: string
      example: apifox-platform-order-001
  responses:
...
    OrderOk:
      description: 成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OrderEnvelope'
...
    XianyuOrderOk:
      description: 成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/XianyuOrderEnvelope'
...
    BotQuoteOk:
      description: 成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BotQuoteEnvelope'
    BotPaymentOk:
      description: 成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BotPaymentEnvelope'
...
    BadRequest:
      description: 参数校验或业务规则失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            message: imageBase64 is required
            data: null
    Unauthorized:
      description: Token 缺失或错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            message: Unauthorized
            data: null
    AuthNotConfigured:
      description: 后端未配置对应接口 Token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            message: API token is not configured
            data: null
  schemas:
    ApiResponseBase:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: ok
    ErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              type: object
              nullable: true
...
    OrderEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/OrderResponse'
...
    XianyuOrderEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/XianyuOrderResponse'
...
    BotQuoteEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/BotQuoteResponse'
    BotPaymentEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/BotPaymentResponse'
...
    QuoteStatus:
      type: string
      enum: [CREATED, FAILED, ORDERED]
    PaymentStatus:
      type: string
      enum: [PENDING, CONFIRMED, REJECTED, USED]
    OrderStatus:
      type: string
      enum: [CREATED, WAIT_SUBMIT, SUBMITTING, SUBMITTED, SUBMIT_FAILED, WAIT_PAY, PAID, TICKETING, ISSUED, FAILED, REFUNDED]
    XianyuFulfillmentStatus:
      type: string
      enum: [WAIT_IMAGE, QUOTED, WAIT_BUYER_PAY, PAID_WAIT_SUBMIT, TICKETING, ISSUED_WAIT_DELIVER, DELIVERY_OUTCOME_PENDING, DELIVERED, NEED_MANUAL, REFUNDED, CLOSED]
...
    MovieTicketInfo:
      type: object
      properties:
        taskId: {type: string}
        provinceName: {type: string}
        cityName: {type: string}
        areaName: {type: string}
        cityCode: {type: string}
        cinemaId: {type: string}
        cinemaCode: {type: string}
        cinemaAddress: {type: string}
        filmId: {type: string}
        filmImg: {type: string}
        customFilmType: {type: integer, format: int32}
        showId: {type: string}
        movieName: {type: string}
        cinemaName: {type: string}
        showTime: {type: string, format: date-time}
        hallName: {type: string}
        planType: {type: string}
        seatCount: {type: integer, format: int32}
        seats:
          type: array
          items: {type: string}
        seatsAndPrice:
          type: object
          additionalProperties: {type: string}
        maxPrice: {type: string}
        totalImagePrice: {type: string}
        imageUrl: {type: string}
    QuoteResponse:
      type: object
      properties:
        quoteNo:
          type: string
        ticketInfo:
          $ref: '#/components/schemas/MovieTicketInfo'
        upstreamPrice:
          type: number
        finalPrice:
          type: number
        totalPrice:
          type: number
        profit:
          type: number
        status:
          $ref: '#/components/schemas/QuoteStatus'
    CreateOrderRequest:
      type: object
      required: [quoteNo, customerId]
      properties:
        quoteNo:
          type: string
        customerId:
          type: string
        paymentNo:
          type: string
    OrderResponse:
      type: object
      properties:
        orderNo: {type: string}
        quoteNo: {type: string}
        customerId: {type: string}
        finalPrice: {type: number}
        totalPrice: {type: number}
        upstreamOrderId: {type: string, nullable: true}
        upstreamOrderNo: {type: string, nullable: true}
        upstreamOrderStatus: {type: integer, format: int32, nullable: true}
        ticketCodeInfo: {type: string, nullable: true}
        submitRetryCount: {type: integer, format: int32}
        lastSubmitError: {type: string, nullable: true}
        lastSyncError: {type: string, nullable: true}
        status:
          $ref: '#/components/schemas/OrderStatus'
        statusText: {type: string}
        terminal: {type: boolean}
        shouldPoll: {type: boolean}
...
    XianyuAmountVerificationSource:
      type: string
      enum: [XIANYU_ORDER_DETAIL_PRICE_INFO_AMOUNT]
...
    XianyuOrderResponse:
      type: object
      properties:
        platformOrderId: {type: string}
        tradeNo: {type: string}
        chatId: {type: string}
        buyerUserId: {type: string}
        itemId: {type: string}
        quoteNo: {type: string, nullable: true}
        localOrderNo: {type: string, nullable: true}
        paidAmount: {type: number, nullable: true}
        quotedAmount: {type: number, nullable: true}
        adjustedAmount: {type: number, nullable: true}
        amountVerificationSource:
          allOf:
            - $ref: '#/components/schemas/XianyuAmountVerificationSource'
          nullable: true
        adjustedAt: {type: string, format: date-time, nullable: true}
        amountVerifiedAt: {type: string, format: date-time, nullable: true}
        status:
          $ref: '#/components/schemas/XianyuFulfillmentStatus'
        lastError: {type: string, nullable: true}
        order:
          type: object
          allOf:
            - $ref: '#/components/schemas/OrderResponse'
          nullable: true
        shouldPoll: {type: boolean}
        shouldDeliver: {type: boolean}
        deliveryMessage: {type: string, nullable: true}
        ticketCodeInfo: {type: string, nullable: true}
        deliveryAttemptId: {type: string, nullable: true}
        deliveryAttemptStartedAt: {type: string, format: date-time, nullable: true}
        deliveryAttemptOutcome:
          type: string
          enum: [SUCCESS, FAILURE]
          nullable: true
        deliveryAttemptChannel: {type: string, nullable: true}
        deliveryAttemptErrorMessage: {type: string, nullable: true}
        deliveryOutcomeRecordedAt: {type: string, format: date-time, nullable: true}
...
    BotImageQuoteRequest:
      type: object
      required: [wechatId, messageId, imageBase64]
      properties:
        wechatId: {type: string}
        nickname: {type: string}
        avatarUrl: {type: string}
        messageId: {type: string}
        imageBase64: {type: string}
    BotQuoteResponse:
      type: object
      properties:
        customerNo: {type: string}
        messageId: {type: string}
        duplicated: {type: boolean}
        latest: {type: boolean}
        latestQuoteNo: {type: string}
        quote:
          $ref: '#/components/schemas/QuoteResponse'
    BotPaymentConfirmRequest:
      type: object
      required: [wechatId, quoteNo, amount, paymentNo]
      properties:
        wechatId: {type: string}
        quoteNo: {type: string}
        amount:
          type: number
          minimum: 0.01
        paymentNo: {type: string}
        proofImageUrl: {type: string}
        confirmer: {type: string}
        remark: {type: string}
    BotPaymentResponse:
      type: object
      properties:
        paymentRecordNo: {type: string}
        customerNo: {type: string}
        quoteNo: {type: string}
        amount: {type: number}
        status:
          $ref: '#/components/schemas/PaymentStatus'
    BotCreateOrderRequest:
      type: object
      required: [wechatId, quoteNo]
      properties:
        wechatId: {type: string}
        quoteNo: {type: string}
        paymentRecordNo:
          type: string
          description: 留空时使用该报价最近一条 CONFIRMED 收款记录